[fs] bio: Need to free integrity payload if the split bio gets memory by itself
Message-id: <1465199742-30937-1-git-send-email-xni@redhat.com>
Patchwork-id: 149150
O-Subject: [RHEL7.3 PATCH BZ#1276454 1/1] Need to free integrity payload if the split bio gets memory by itself
Bugzilla: 1276454
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

RHEL7-only patch, not upstream.

Upstream moved to a bio_split implementation that actually clones the
bio into two smaller bios.  Because those two smaller bios are now
allocated using bio_alloc, they are freed on the last put, and any
integrity payload that was allocated for them is also freed at that
time.  We can't backport that implementation, because it depends on
immutable biovecs (which break kabi) to be performant (immutable biovecs
make bio cloning much faster).

bio1 and bio2 (the split bios) are embedded inside of the bio_pair
structure, which means they are not allocated via bio_alloc, and of
course do not need to be freed.  Because bio integrity data is freed
when the bio is freed, and we've now introduced two bios that will not
be freed, this means that the integrity data for those two bios will
never be freed.  This patch fixes that by doing the freeing in the
bio_split code itself.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1276454
Brew:     https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11136481
Testing:  The repoducer steps in this bug
Signed-off-by: Rafael Aquini <aquini@redhat.com>