[netdrv] qed: Add VF->PF channel infrastructure
Message-id: <1465316705-7370-83-git-send-email-hpatil@redhat.com>
Patchwork-id: 149505
O-Subject: [PATCH v2 RHEL 7.3 082/107] qed: Add VF->PF channel infrastructure
Bugzilla: 1275807 1275811
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Yuval Mintz <Yuval.Mintz@qlogic.com>

BZ 1275807

net commit id 37bff2b9c6addf6216c8d04e95be596678e8deff

Communication between VF and PF is based on a dedicated HW channel;
VF will prepare a messge, and by signaling the HW the PF would get a
notification of that message existance. The PF would then copy the
message, process it and DMA an answer back to the VF as a response.

The messages themselves are TLV-based - allowing easier backward/forward
compatibility.

This patch adds the infrastructure of the channel on the PF side -
starting with the arrival of the notification and ending with DMAing
the response back to the VF.

It also adds a dummy-response as reference, as it only lays the
groundwork of the communication; it doesn't really add support of any
actual messages.

Signed-off-by: Yuval Mintz <Yuval.Mintz@qlogic.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Rafael Aquini <aquini@redhat.com>