[netdrv] qlcnic: constify qlcnic_mbx_ops structure
Message-id: <1465323903-13007-15-git-send-email-hpatil@redhat.com>
Patchwork-id: 149535
O-Subject: [PATCH RHEL 7.3 14/23] qlcnic: constify qlcnic_mbx_ops structure
Bugzilla: 1275799
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Julia Lawall <julia.lawall@lip6.fr>

BZ 1275799

net commit id 571f2c11b343cd6997f35a21f6caa0f78e87fc84

The only instance of a qlcnic_mbx_ops structure is never modified.  Thus
the declaration of the structure and all references to the structure type
can be made const.

In the definition of the qlcnic_mailbox structure, the ops field is no
longer lined up with the other fields.  This was left as is, to avoid a lot
of trivial changes on the other lines.

Done with the help of Coccinelle.

Signed-off-by: Julia Lawall <Julia.Lawall@lip6.fr>
Acked-by: Sony Chacko <sony.chacko@qlogic.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Rafael Aquini <aquini@redhat.com>