[scsi] bnx2fc: Check sc_cmd device and host pointer before returning the command to the mid-layer
Message-id: <1465380196-6762-13-git-send-email-mlombard@redhat.com>
Patchwork-id: 149774
O-Subject: [RHEL7.3 e-stor PATCH 12/13] bnx2fc: Check sc_cmd device and host pointer before returning the command to the mid-layer.
Bugzilla: 1273084
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chad Dupuis <chad.dupuis@qlogic.com>

When we are in connection recovery and the internal command timer on a
request pops, either the scsi_cmnd->device or scsi_cmnd->device->host
back pointers may be NULL as the device that the command that the
request was submitted on may have been subsequently reaped due to the
connection recovery. This can cause one or both of the pointers above to
be NULL and cause a system crash if we try to return the command to the
midlayer.

Instead, double check the pointers before the return to the midlayer so
as to prevent the crash and let the upper layers finish the session
recovery and rediscover the device.

Signed-off-by: Chad Dupuis <chad.dupuis@qlogic.com>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit d02327a863168647b9e6fde610c4730ff4837f9e)
Signed-off-by: Rafael Aquini <aquini@redhat.com>