[infiniband] ib/iser: Remove dead code in fmr_pool alloc/free
Message-id: <1465597803-11831-274-git-send-email-ddutile@redhat.com>
Patchwork-id: 150966
O-Subject: [PATCH RHEL-7.3 0273/1003] IB/iser: Remove dead code in fmr_pool alloc/free
Bugzilla: 1169955 1259940 1262728 1275187 1275209 1275423 1275425 1289615 1291874 1292284 1292872 1296195 1296269 1296338 1296344 1298707 1302166 1310156
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Laurence Oberman <loberman@redhat.com>

From: Sagi Grimberg <sagig@mellanox.com>

In the past the we always tried to allocate an fmr_pool
and if it failed on ENOSYS (not supported) then we continued
with dma mr. This is not the case anymore and if we tried to
allocate an fmr_pool then it is supported and we expect to succeed.

Also, the check if fmr_pool is allocated when free is called is
redundant as well as we are guaranteed it exists.

Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 8c18ed03a95cb6c3543b0a9e0df5e9366baea5df)

Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>