[infiniband] ib/srp: Allow activation of scsi-mq for SRP in driver
Message-id: <1465597803-11831-568-git-send-email-ddutile@redhat.com>
Patchwork-id: 151104
O-Subject: [PATCH RHEL-7.3 0567/1003] IB/srp: Allow activation of scsi-mq for SRP in driver
Bugzilla: 1169955 1259940 1262728 1275187 1275209 1275423 1275425 1289615 1291874 1292284 1292872 1296195 1296269 1296338 1296344 1298707 1302166 1310156
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Laurence Oberman <loberman@redhat.com>

Allow the activation of the scsi-mq feature in the SRP driver as opposed
to the current existing mechanism which activates it globally for all SCSI
hosts.  Activating it globally remains tech preview in RHEL7.3.

Based upon an earlier patch by Johannes Thumshirn.

This patch is RHEL-specific.

Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>