[fs] libceph: pick a different monitor when reconnecting
Message-id: <1465829862-17784-85-git-send-email-idryomov@redhat.com>
Patchwork-id: 151510
O-Subject: [RHEL7.3 PATCH 084/173] libceph: pick a different monitor when reconnecting
Bugzilla: 1344930
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>

From: Ilya Dryomov <idryomov@gmail.com>

Don't try to reconnect to the same monitor when we fail to establish
a session within a timeout or it's lost.

For that, pick_new_mon() needs to see the old value of cur_mon, so
don't clear it in __close_session() - all calls to __close_session()
but one are followed by __open_session() anyway.  __open_session() is
only called when a new session needs to be established, so the "already
open?" branch, which is now in the way, is simply dropped.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 0e04dc26cc594d31ee6b1382b452b6bc83b57937)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>