[redhat] config settings and minor fixes for RHEL7
Message-id: <1465909743-5781-1-git-send-email-prarit@redhat.com>
Patchwork-id: 151783
O-Subject: [RHEL7.3 PATCH BZ 1339663 04/91 v2] config settings and minor fixes for RHEL7
Bugzilla: 1339663
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1339663

The previous two patches will not compile without the following additional
changes.  The config options for GPIOLIB & PINCTRL_BAYTRAIL must be set,
and PINCTRL should be enabled via LPSS for RHEL7.  (Note that this patch
is taken from linux commit 0f53143 ("x86/intel/lpss: Add pin control
support to Intel low power subsystem")).

When PINCTRL is enabled, there is a minor ssize_t file operations
conflict that is also resolved in this patch.

There are two KABI changes, the __GENKSYMS__ wrapping of seq_file.h, and
the moving of struct dev_pin_info to struct rh_device.

[v2]: A conflict was introduced since this code was built and submitted,
and the time that the code was applied to the tree by raquini, the
maintainer.  The drm code has introduced simple gpio stub code which can
can be removed with this patch that enables GPIOLIB.  Note that the gpio
code in RHEL7 is still older than upstream which results in the change in
intel_dsi_init() to remove the flags parameter.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: rclark@redhat.com
Cc: raquini@redhat.com
Cc: dzickus@redhat.com
Cc: jmcnicol@redhat.com
Cc: linville@redhat.com
Cc: tcamuso@redhat.com
Signed-off-by: Rafael Aquini <aquini@redhat.com>