[include] ib/core: Support accessing SA in virtualized environment
Message-id: <1465917990-1132-87-git-send-email-kheib@redhat.com>
Patchwork-id: 151916
O-Subject: [RHEL7.3 PATCH 086/181] IB/core: Support accessing SA in virtualized environment
Bugzilla: 1275159 1296272 1296405 1298421 1298422 1298423 1298424 1298425
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Eli Cohen <eli@mellanox.com>

Per the ongoing standardisation process, when virtual HCAs are present
in a network, traffic is routed based on a destination GID. In order to
access the SA we use the well known SA GID.

We also add a GRH required boolean field to the port attributes which is
used to report to the verbs consumer whether this port is connected to a
virtual network. We use this field to realize whether we need to create
an address vector with GRH to access the subnet administrator. We clear
the port attributes struct before calling the hardware driver to make
sure the default remains that GRH is not required.

Signed-off-by: Eli Cohen <eli@mellanox.com>
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit a0c1b2a3508714281f604db818fa0cc83c2f9b6a)
Signed-off-by: Kamal Heib <kheib@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>