[infiniband] ib/mlx4: Fix unaligned access in send_reply_to_slave
Message-id: <1465917990-1132-154-git-send-email-kheib@redhat.com>
Patchwork-id: 151921
O-Subject: [RHEL7.3 PATCH 153/181] IB/mlx4: Fix unaligned access in send_reply_to_slave
Bugzilla: 1275159 1296272 1296405 1298421 1298422 1298423 1298424 1298425
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: shamir rabinovitch <shamir.rabinovitch@oracle.com>

The problem is that the function 'send_reply_to_slave' gets the
'req_sa_mad' as a pointer whose address is only aliged to 4 bytes
but is 8 bytes in size.  This can result in unaligned access faults
on certain architectures.

Sowmini Varadhan pointed to this reply from Dave Miller that say
that memcpy should not be used to solve alignment issues:
https://lkml.org/lkml/2015/10/21/352

Optimization of memcpy to 'ldx' instruction can only happen if the
compiler knows that the size of the data we are copying is 8 bytes
and it assumes it is aligned to 8 bytes. If the compiler know the
type is not aligned to 8 it must not optimize the 8 byte copy.
Defining the data type as aligned to 4 forces the compiler to treat
all accesses as though they aren't aligned and avoids the 'ldx'
optimization.

Full credit for the idea goes to Jason Gunthorpe
<jgunthorpe@obsidianresearch.com>.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 04ef0f1a0169a14b8e653af1178524ab4390133f)
Signed-off-by: Kamal Heib <kheib@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>