[kernel] sched/deadline: Disable SCHED_DEADLINE programmatically
Message-id: <1465932845-8308-5-git-send-email-xlpang@redhat.com>
Patchwork-id: 152656
O-Subject: [RHEL7.3 PATCH v3 103/100] sched/deadline: Disable SCHED_DEADLINE programmatically
Bugzilla: 1298387
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Resolves:
https://bugzilla.redhat.com/show_bug.cgi?id=1298387

Brew build:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11178233

This patch is RHEL-only.

According to PM's intentions, SCHED_DEADLINE code is disabled in RHEL7
and will be enabled only in the RT kernel which is based on some version
of RHEL7 after this.

So this patch disables it in the code level. SCHED_DEADLINE code still
gets compiled during regular kernel builds after this patch, it can help
catch issues in the scheduler code and subsequent merges before it gets
to RT kernel.

We also can easily back out this patch if SCHED_DEADLINE is needed in
future RHEL7.

Signed-off-by: Xunlei Pang <xlpang@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>