[infiniband] ib/rdmavt: Create module framework and handle driver registration
Message-id: <20160614193608.23159.24825.stgit@dhcp187-101.khw.lab.eng.bos.redhat.com>
Patchwork-id: 152806
O-Subject: [RHEL-7.3 BZs 1273170 1272062 PATCH 032/431] IB/rdmavt: Create module framework and handle driver registration
Bugzilla: 1272062 1273170
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Dennis Dalessandro <dennis.dalessandro@intel.com>

Bugzilla: 1273170
Upstream status: (imported from commit 0194621b225348428c212f330c26d194fc77bd15)

This patch introduces the basics for a new module called rdma_vt. This new
driver is a software implementation of the InfiniBand verbs and aims to
replace the multiple implementations that exist and duplicate each others'
code.

While the call to actually register the device with the IB core happens in
rdma_vt, most of the work is still done in the drivers themselves. This
will be changing in a follow on patch this is just laying the groundwork
for this infrastructure.

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Alex Estrin <aestrin@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>