[infiniband] ib/qib: Remove driver specific members from qib qp type
Message-id: <20160614193654.23159.2270.stgit@dhcp187-101.khw.lab.eng.bos.redhat.com>
Patchwork-id: 152746
O-Subject: [RHEL-7.3 BZs 1273170 1272062 PATCH 073/431] IB/qib: Remove driver specific members from qib qp type
Bugzilla: 1272062 1273170
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Dennis Dalessandro <dennis.dalessandro@intel.com>

Bugzilla: 1273170  1272062
Upstream: (imported from commit ffc269075bce7dd895d1fc8eca6367573df0ddbc)

In preparation for moving the queue pair data structure to rdmavt the
members of the driver specific queue pairs which are not common need to be
pushed off to a private driver structure. This structure will be available
in the queue pair once moved to rdmavt as a void pointer. This patch while
not adding a lot of value in and of itself is a prerequisite to move the
queue pair out of the drivers and into rdmavt.

The driver specific, private queue pair data structure should condense as
more of the send side code moves to rdmavt.

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>