[infiniband] ib/rdmavt: Add post send to rdmavt
Message-id: <20160614193719.23159.78037.stgit@dhcp187-101.khw.lab.eng.bos.redhat.com>
Patchwork-id: 152707
O-Subject: [RHEL-7.3 BZs 1273170 1272062 PATCH 095/431] IB/rdmavt: Add post send to rdmavt
Bugzilla: 1272062 1273170
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Dennis Dalessandro <dennis.dalessandro@intel.com>

Bugzilla: 1273170
Upstream: (imported from commit bfbac097b6e8023e10fdadab2527d0a1a3160d7e)

Add in a post_send and post_one_send to rdmavt. The ULP will provide a WQE
to rdmavt which will then walk and queue each element. Rdmavt will then
queue the work to be done in the driver or kick the driver's progress
routine.

There needs to be a follow on patch which adds in another lock for the
head of the queue so that it can be added to and read from in parallel.
This will touch protocol handlers and require other changes in the
drivers. This will be done separately.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>