[infiniband] rdma/hfi1: Tune for unknown channel if configuration file is absent
Message-id: <20160614193949.23159.26815.stgit@dhcp187-101.khw.lab.eng.bos.redhat.com>
Patchwork-id: 153041
O-Subject: [RHEL-7.3 BZs 1273170 1272062 PATCH 226/431] staging/rdma/hfi1: Tune for unknown channel if configuration file is absent
Bugzilla: 1272062 1273170
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Easwar Hariharan <easwar.hariharan@intel.com>

Bugzilla: 1273170
Upstream: (imported from commit 97167e8134150eb5104e19fd7208e3ac3525f48b)

Currently, the driver fails to tune the SerDes and therefore prevents
link up if the configuration file is missing or fails parsing or
validation. This patch adds a fallback option so that the 8051 is asked
to tune for an unknown channel and possibly get the link up if tuning
succeeds. It also adds a user-friendly message to update the
configuration file if it is out-of-date.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Signed-off-by: Jubin John <jubin.john@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>