[infiniband] ib/hfi1: Improve performance of interval RB trees
Message-id: <20160614194301.23159.21096.stgit@dhcp187-101.khw.lab.eng.bos.redhat.com>
Patchwork-id: 152932
O-Subject: [RHEL-7.3 BZs 1273170 1272062 PATCH 395/431] IB/hfi1: Improve performance of interval RB trees
Bugzilla: 1272062 1273170
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Mitko Haralanov <mitko.haralanov@intel.com>

Bugzilla: 1273170
Upstream: (imported from commit 67caea1fec85266590dddfa431e1cb000bc942f4)

store user-specific callback for the various tree operations.
These handlers are put on a doubly-linked list. When a RB
tree function is called, the list is searched for the handler
of the particular tree.

The list which holds the handlers is modified very rarely - when
a handler is created and when a handler is removed. On the other
hand, it is searched very often. This a perfect usage scenario
for RCU.

The result is a much lower overhead of traversing the list as most
of the time no locking will be required.

Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>