[sound] alsa: regmap: regcache: allow read-only regs to be cached
Message-id: <146616708042-4-patchset-jkysela@redhat.com>
Patchwork-id: 153528
O-Subject: [RHEL 7.3 PATCH 03/04] ALSA/regmap: regmap: regcache: allow read-only regs to be cached
Bugzilla: 1285520
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The regmap_writeable() check should not be done in
regcache_write() because this prevents read-only
registers to be cached. After a read on a read-only
register its value will not be stored in the cache
and the next time someone will try to read it the
value will be read from the bus instead of the
cache.

Instead the regmap_writeable() check should be done
in _regmap_write() to prevent callers from writing
to read-only registers.

Signed-off-by: Ionut Nicu <ioan.nicu.ext@nsn.com>
Signed-off-by: Mark Brown <broonie@linaro.org>

(picked from kernel: 515f2261703d09c6b647a5687b7d657dd5911065)
rhbz#1285520
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>